+++ packages/terra-button/src/terra-dev-site/test/button/CompactButton.test.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import Button from '../../../Button';
-
-const CompactButton = () => <Button id="compactButton" text="Compact" isCompact />;
-
-export default CompactButton;